lambda_report/pyproject.toml
[tool.poetry]
name = "temp"
version = "0.1.0"
description = ""
authors = ["pirocheto <pirocheto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fpdf2 = "^2.7.8"
matplotlib = "^3.8.3"
pandas = "^2.2.1"
boto3 = "^1.34.58"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.8.3"
pandas = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
